Move non-ticket actions to orders/actions.py
