+++ orders/templates/orders/order_created.slack
