+++ tickets/tests/test_views.py
@@ -6,6 +6,7 @@ from ironcage.tests import utils
 
 from . import factories
 
+from orders import actions as order_actions
 from tickets import actions
 from tickets.models import TicketInvitation
 
@@ -351,7 +352,7 @@ class OrderTests(TestCase):
         ticket = order.all_tickets()[-1]
 
         with utils.patched_refund_creation():
-            actions.refund_ticket(ticket, 'Refund requested by user')
+            order_actions.refund_ticket(ticket, 'Refund requested by user')
 
         self.client.force_login(order.purchaser)
         rsp = self.client.get(f'/orders/{order.order_id}/', follow=True)
@@ -589,7 +590,7 @@ class RefundCreditNoteTests(TestCase):
         ticket = cls.order.all_tickets()[0]
 
         with utils.patched_refund_creation():
-            actions.refund_ticket(ticket, 'Refund requested by user')
+            order_actions.refund_ticket(ticket, 'Refund requested by user')
 
         cls.refund = cls.order.refunds.get()
